refactor(calendar): split calendar.rs into a cohesive module folder
The single calendar.rs had grown to 1670 lines (≈890 prod + ≈780 test),
mixing iCal serialization, CalDAV transport, event processing, and the
Tauri command surface in one file.

Split it into integrations/calendar/ following the existing daylite/
folder convention, by concern:

- types.rs    public Specta types + internal RawVEvent/PendingEvent
- ical.rs     RFC 5545 payload building and iCal parsing
- caldav.rs   CalDAV REPORT fetch, response parsing, and write cores
- events.rs   classification, project resolution, ordering, absence mapping
- commands.rs the four Tauri commands (orchestration only)
- mod.rs      wiring + re-exports keeping integrations::calendar::* stable

Pure code move: no behavior change. Each test was relocated next to the
code it covers; all 38 tests are preserved. The public command paths used
by lib.rs are unchanged via re-exports.
