+++ src-tauri/src/integrations/calendar/mod.rs
@@ -1,0 +1,8 @@
+mod caldav;
+mod commands;
+mod events;
+mod ical;
+mod types;
+
+pub use commands::{create_assignment, delete_assignment, load_week_events, update_assignment};
+pub use types::{CalendarCellEvent, CalendarEventKind, EmployeeWeekEvents};
